+++ .github/workflows/run-e2e-tests-node20-wcon.yaml
@@ -1,4 +1,4 @@
-name: RUN_E2E_TESTS_NODE16_WCON
+name: RUN_E2E_TESTS_NODE20_WCON
 on:
   push:
     branches:
@@ -12,9 +12,9 @@ on:
       - '**.md'
 
 env:
-  AZURE_FUNCTIONAPP_NAME: gae-fa-node16-wcon
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/node16'
-  NODE_VERSION: '16.x'
+  AZURE_FUNCTIONAPP_NAME: gae-fa-node20-wcon
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/node20'
+  NODE_VERSION: '20.x'
 
 jobs:
   run:
@@ -24,10 +24,10 @@ jobs:
     - name: Checkout
       uses: actions/checkout@v2
 
-    - name: Set Node.js 16.x for GitHub Action
+    - name: Set Node.js 20.x for GitHub Action
       uses: actions/setup-node@v1
       with:
-        node-version: 16.x
+        node-version: 20.x
 
     - name: Run Npm Install for GitHub Action
       run: npm install
@@ -53,7 +53,7 @@ jobs:
     - name: 'Login via Azure CLI'
       uses: azure/login@v1
       with:
-        creds: ${{ secrets.RBAC_GAE_FA_NODE16_WCON }}
+        creds: ${{ secrets.RBAC_GAE_FA_NODE20_WCON }}
 
     - name: E2E Run Azure Functions Action
       uses: ./
