+++ action.yml
@@ -55,5 +55,5 @@ branding:
   icon: 'functionapp.svg'
   color: 'blue'
 runs:
-  using: 'node16'
+  using: 'node20'
   main: 'lib/main.js'
