--- tests/e2e/node16/.vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-    "recommendations": [
-        "ms-azuretools.vscode-azurefunctions"
-    ]
-}
