--- tests/e2e/node16/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "node16",
-  "version": "1.0.0",
-  "description": "E2E Test for GitHub action",
-  "scripts": {
-    "test": "echo \"No tests yet...\""
-  },
-  "author": "",
-  "dependencies": {
-    "lodash": "^4.17.21"
-  }
-}
