+++ tests/e2e/node20/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "node20",
+  "version": "1.0.0",
+  "description": "E2E Test for GitHub action",
+  "scripts": {
+    "test": "echo \"No tests yet...\""
+  },
+  "author": "",
+  "dependencies": {
+    "lodash": "^4.17.21"
+  }
+}
