+++ tests/e2e/node20/sha.txt
@@ -1,0 +1,1 @@
+sha-txt-node20-placeholder
